changed interpolation to Float for memory conservation

+++ src/main/java/smptk/image/BSplineCoefficients.java
@@ -50,16 +50,16 @@ package smptk.image;
 public class BSplineCoefficients {
 
 	
-	private static double getInitialAntiCausalCoefficientMirrorOnBounds(double[] c,
-			double z, double tolerance) {
-		return ((z * c[c.length - 2] + c[c.length - 1]) * z / (z * z - 1.0));
+	private static float getInitialAntiCausalCoefficientMirrorOnBounds(float[] c,
+			float z, float tolerance) {
+		return (float) ((z * c[c.length - 2] + c[c.length - 1]) * z / (z * z - 1.0));
 	} /* end getInitialAntiCausalCoefficientMirrorOnBounds */
 
 	/*------------------------------------------------------------------*/
-	static double getInitialCausalCoefficientMirrorOnBounds(double[] c, double z,
-			double tolerance) {
-		double z1 = z, zn = Math.pow(z, c.length - 1);
-		double sum = c[0] + zn * c[c.length - 1];
+	static float getInitialCausalCoefficientMirrorOnBounds(float[] c, float z,
+			float tolerance) {
+		float z1 = z, zn = (float) Math.pow(z, c.length - 1);
+		float sum = c[0] + zn * c[c.length - 1];
 		int horizon = c.length;
 
 		if (0.0 < tolerance) {
@@ -72,39 +72,39 @@ public class BSplineCoefficients {
 			sum = sum + (z1 + zn) * c[n];
 			z1 = z1 * z;
 		}
-		return (sum / (1.0 - Math.pow(z, 2 * c.length - 2)));
+		return (float) (sum / (1.0 - Math.pow(z, 2 * c.length - 2)));
 	} /* end getInitialCausalCoefficientMirrorOnBounds */
 
 
 	
-	static public void getSplineInterpolationCoefficients(int degree, double[] c) throws Exception {
+	static public void getSplineInterpolationCoefficients(int degree, float[] c) throws Exception {
 		
-		double z[] = null;
+		float z[] = null;
 		if (degree == 0 || degree == 1) { 
-			z = new double[0];			
+			z = new float[0];			
 		}
 		else if (degree == 2) { 
-			z = new double[1];
-			z[0] = Math.sqrt(8.0) - 3.0;
+			z = new float[1];
+			z[0] = (float)(Math.sqrt(8.0) - 3.0);
 		}
 		else if(degree == 3) { 
-			z = new double[1];
-			z[0] = Math.sqrt(3.0) - 2.0;
+			z = new float[1];
+			z[0] =(float) (Math.sqrt(3.0) - 2.0);
 		}
 		else {
 			throw new Exception("SplineOrder must be between 0 and 3");
 		}
 		
-		double tolerance = 0;//1.;//  Float.intBitsToFloat((int)0x33FFFFFF);
+		float tolerance = 0;//1.;//  Float.intBitsToFloat((int)0x33FFFFFF);
 		
 
-		double lambda = 1.0;
+		float lambda = 1;
 
 		if (c.length == 1) {
 			return;
 		}
 		for (int k = 0; (k < z.length); k++) {
-			lambda = lambda * (1.0 - z[k]) * (1.0 - 1.0 / z[k]);
+			lambda = lambda * (1 - z[k]) * (1 - 1 / z[k]);
 		}
 		for (int n = 0; (n < c.length); n++) {
 			c[n] = c[n] * lambda;
